requirements.txt
colored==2.3.1
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.2
h5py==3.14.0
hdf5plugin==6.0.0
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
nexusformat==2.0.1
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pillow==11.3.0
Pygments==2.19.2
pyparsing==3.3.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.13.1
six==1.17.0
tzdata==2025.3
xarray==2024.7.0
zipp==3.23.0
